Move state flow tests to their own file
